Module Multipart_form.Content_type

module Type : sig ... end

module Subtype : sig ... end

module Parameters : sig ... end

type t = {

ty : Type.t;

subty : Subtype.t;

parameters : (string * Parameters.value) list;

}

val make : Type.t -> Subtype.t -> Parameters.value Parameters.Map.t -> t

val equal : t -> t -> bool

val pp : t Fmt.t

val of_string : string -> ( t, [> `Msg of string ] ) Stdlib.result

of_string str returns the Content-Type value from a string which come from your HTTP stack. NOTE: the string must finish with "\r\n". If you are not sure about the value returned by your HTTP stack, you should append it.

val to_string : t -> string